Add single-header generator and refactor naming conventions
Introduce generate_header.py to auto-generate ctui_c.h as a single-header library, integrated as a pre-build step. Refactor codebase to use snake_case for functions and variables. Add ctui_container_screen.cpp and update Screen class construction and color handling. Update project files and filters, and improve type safety in Percent.

+++ CroissanTUI/src/ctui_pct.cpp
@@ -5,8 +5,8 @@ namespace ctui {
 	Percent::Percent(long double val) : pct(val) {}
 
 	int Percent::to_pixels(int max_width) const {
-		return max_width * this->pct;
+		return static_cast<int>(max_width * this->pct);
 	}
-	Percent operator""_pct(unsigned long long val) { return Percent(double(val) / 100.0); }
+	Percent operator""_pct(unsigned long long val) { return Percent(static_cast<double>(val) / 100.0); }
 
 }
